common method (_run_cmd) moved to plugin base class
