+++ console/cocos2d_compile.py
@@ -13,7 +13,6 @@
 __docformat__ = 'restructuredtext'
 
 import sys
-import subprocess
 import os
 import json
 import inspect
@@ -48,12 +47,6 @@ class CCPluginCompile(cocos2d.CCPlugin):
             return None
 
         return project_dir
-
-    def _run_cmd(self, command):
-        print "\ncommand:"+command+"\n"
-        ret = subprocess.call(command, shell=True)
-        if ret != 0:
-            raise Exception("Error running command")
 
     def build_android(self):
         project_dir = self._build_project_dir('proj.android', 'Android')
